backend/main.py
from fastapi import FastAPI, UploadFile, File, Body
from pydantic import BaseModel
from typing import List
import uuid
import docx
import PyPDF2
import numpy as np
import openai  # pip install openai

app = FastAPI()

documents = {}
sessions = {}

@app.get("/health")
def health():
    return {"status": "ok"}

class AskRequest(BaseModel):
    session_id: str
    document_ids: List[str]
    question: str

def cosine_similarity(vec1, vec2):
    return float(np.dot(vec1, vec2) / (np.linalg.norm(vec1) * np.linalg.norm(vec2)))

def embed_text(text: str):
    response = openai.Embedding.create(
        model="text-embedding-ada-002",
        input=text
    )
    return response['data'][0]['embedding']

@app.post("/ask")
def ask(request: AskRequest = Body(...)):
    if request.session_id not in sessions:
        sessions[request.session_id] = []

    all_chunks = []
    for doc_id in request.document_ids:
        doc_chunks = documents.get(doc_id, {}).get("chunks", [])
        for chunk in doc_chunks:
            chunk["document_id"] = doc_id
            all_chunks.append(chunk)

    question_embedding = embed_text(request.question)

    similarities = []
    for chunk in all_chunks:
        score = cosine_similarity(question_embedding, chunk["embedding"])
        similarities.append((score, chunk))

    similarities.sort(key=lambda x: x[0], reverse=True)
    relevant_chunks = [chunk for _, chunk in similarities[:5]]

    answer = " ".join([chunk["text"] for chunk in relevant_chunks])

    sessions[request.session_id].append({"role": "user", "content": request.question})
    sessions[request.session_id].append({"role": "assistant", "content": answer})

    safe_chunks = [
        {
            "document_id": str(chunk["document_id"]),
            "chunk_id": int(chunk["chunk_id"]),
            "text": chunk["text"]
        }
        for chunk in relevant_chunks
    ]

    return {
        "answer": answer,
        "session_id": request.session_id,
        "source_chunks": safe_chunks,
        "batch_size": len(safe_chunks),
        "tokens_used": {
            "prompt_tokens": 10,
            "candidates_tokens": 20,
            "total_tokens": 30
        }
    }

@app.post("/upload")
async def upload(file: UploadFile = File(...)):
    doc_id = str(uuid.uuid4())

    text_content = ""
    if file.filename.endswith(".txt"):
        raw_bytes = await file.read()
        try:
            text_content = raw_bytes.decode("utf-8")
        except UnicodeDecodeError:
            text_content = raw_bytes.decode("latin-1")
    elif file.filename.endswith(".pdf"):
        reader = PyPDF2.PdfReader(file.file)
        text_content = "\n".join([page.extract_text() or "" for page in reader.pages])
    elif file.filename.endswith(".docx"):
        doc = docx.Document(file.file)
        text_content = "\n".join([para.text for para in doc.paragraphs])
    else:
        text_content = "Unsupported file type."

    chunks = []
    for i in range(0, len(text_content), 500):
        chunk_text = text_content[i:i+500]
        chunk_embedding = embed_text(chunk_text)
        chunks.append({
            "chunk_id": i // 500,
            "text": chunk_text,
            "embedding": chunk_embedding
        })

    documents[doc_id] = {
        "filename": file.filename,
        "status": "completed",
        "chunks": chunks
    }

    return {
        "document_id": doc_id,
        "filename": file.filename,
        "message": "Document processed successfully."
    }

# Make sure embed_text is accessible for tests
__all__ = ["app", "embed_text"]
