backend/requirements.txt
fastapi
uvicorn
python-docx
PyPDF2
python-multipart
